services/import-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
